LoongArch: Support new relocation types in v2.00 ABI
After LoongArch upgraded[1] the EFI ABI document from 1.0 to 2.0[2], a new set of
relocation types (starting from number 64) was added to replace the 1.0 abi
relocation. It is implemented in binutils 2.40 and gcc 13.

This patch adds support for abi 2.0 relocation, and plans to remove the code
related to 1.0 relocation in future versions.

[1]: https://github.com/loongson/LoongArch-Documentation/pull/57
[2]: https://loongson.github.io/LoongArch-Documentation/LoongArch-ELF-ABI-EN.html#_appendix_revision_history

Signed-off-by: Xiaotian Wu <wuxiaotian@loongson.cn>